server:  Drop it and also drop the client
After a while of working on this I have decided to drop this.
The option I'm going with is to use a service like tailscale to
have network access to all the devices  which will allow me
to have full access to a new API I'll be working on.  This will
also allow me to expand out the project to allow for multi ways
to get data and send data to the API.